+++ src/main/java/com/fpt/gsu25se47/schoolpsychology/dto/request/ReportCategoryRequest.java
@@ -2,9 +2,11 @@ package com.fpt.gsu25se47.schoolpsychology.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ReportCategoryRequest {
 
     private Integer categoryId;
-    private Integer score;
+    private BigDecimal score;
 }
